requirements.txt
pip~=24.2
attrs~=23.2.0
wheel~=0.44.0
cryptography~=43.0.1
pillow~=10.4.0
Jinja2~=3.1.4
base~=0.0.0
certifi~=2024.6.2
fastapi~=0.99.1
uvicorn~=0.30.6
requests~=2.32.3
aiofiles~=23.2.1
aiohttp~=3.9.5
typing_extensions~=4.12.2
urllib3~=2.2.2
idna~=3.7
python-dotenv~=1.0.1